Somando dois números

+++ exe003.py
@@ -1,4 +1,4 @@
-n1 = int(input("Digite um valor: "))
-n2 = int(input("Digite outro valor: "))
+n1 = float(input("Digite um valor: "))
+n2 = float(input("Digite outro valor: "))
 
 print(n1 + n2)
